refactor some functions in views to trait and add defautlConfig for ViewClass
